Convert `dt::shared_lock` class into template
