+++ c/extras/aggregator.h
@@ -15,108 +15,6 @@
 #include "utils.h"
 
 
-/**
- * Shared mutex implementation with busy `while` loops instead of `std::mutex`.
- */
-class shared_mutex {
-  private:
-    size_t state;
-    static constexpr size_t EXCLMASK = 1ULL << (sizeof(size_t) * 8 - 1);
-
-  public:
-    shared_mutex() : state(0) {}
-
-    void lock_shared() {
-      size_t state_old;
-
-      while (true) {
-        #pragma omp atomic read
-        state_old = state;
-        // This check is required to prevent deadlocks, so that exclusive `lock()`
-        // takes priority over the `lock_shared()` when competing. Note, that
-        // the `state` should be "omp atomic read", otherwise behavior is unpredictable
-        // on some systems.
-        if (state_old & EXCLMASK) continue;
-
-        #pragma omp atomic capture
-        {state_old = state; ++state;}
-
-        if (state_old & EXCLMASK) {
-          #pragma omp atomic update
-          --state;
-        } else break;
-      }
-    }
-
-    void unlock_shared() {
-      #pragma omp atomic update
-      --state;
-    }
-
-    void lock() {
-      size_t state_old;
-      bool exclusive_request = false;
-
-      while (true) {
-        #pragma omp atomic capture
-        {state_old = state; state |= EXCLMASK;}
-
-        if ((state_old & EXCLMASK) && !exclusive_request) continue;
-        if (state_old == EXCLMASK) break;
-        exclusive_request = true;
-      }
-    }
-
-    void unlock() {
-      #pragma omp atomic update
-      state &= ~EXCLMASK;
-    }
-};
-
-
-class shared_lock {
-  private:
-    shared_mutex& mutex;
-    bool exclusive;
-    uint64_t : 56;
-
-  public:
-    shared_lock(shared_mutex& m, bool excl = false)
-      : mutex(m), exclusive(excl)
-    {
-      if (exclusive) {
-        mutex.lock();
-      } else {
-        mutex.lock_shared();
-      }
-    }
-
-    ~shared_lock() {
-      if (exclusive) {
-        mutex.unlock();
-      } else {
-        mutex.unlock_shared();
-      }
-    }
-
-    void exclusive_start() {
-      if (!exclusive) {
-        mutex.unlock_shared();
-        mutex.lock();
-        exclusive = true;
-      }
-    }
-
-    void exclusive_end() {
-      if (exclusive) {
-        mutex.unlock();
-        mutex.lock_shared();
-        exclusive = false;
-      }
-    }
-};
-
-
 typedef std::unique_ptr<double[]> DoublePtr;
 struct ex {
   int64_t id;
