+++ c/read/parallel_reader.cc
@@ -287,7 +287,7 @@ void ParallelReader::realloc_output_columns(size_t ichunk, size_t new_alloc)
   nrows_allocated = new_alloc;
   g.trace("Too few rows allocated, reallocating to %zu rows", nrows_allocated);
 
-  dt::shared_lock lock(shmutex, /* exclusive = */ true);
+  dt::shared_lock<dt::shared_mutex> lock(shmutex, /* exclusive = */ true);
   g.columns.set_nrows(nrows_allocated);
 }
